+++ packages/scrabble-solver/src/components/Sidebar/Sidebar.tsx
@@ -4,7 +4,7 @@ import React, { FunctionComponent, ReactNode } from 'react';
 import { cross } from 'icons';
 import { useTranslate } from 'state';
 
-import IconButton from '../IconButton';
+import Button from '../Button';
 
 import { Section } from './components';
 import styles from './Sidebar.module.scss';
@@ -30,7 +30,7 @@ const Sidebar: FunctionComponent<Props> = ({ children, className, hidden, title,
       <div className={styles.header}>
         <h1 className={styles.title}>{title}</h1>
 
-        <IconButton
+        <Button
           className={styles.closeButton}
           disabled={hidden}
           icon={cross}
