+++ packages/scrabble-solver/src/components/index.ts
@@ -1,8 +1,8 @@
 export { default as Board } from './Board';
+export { default as Button } from './Button';
 export { default as Checkbox } from './Checkbox';
 export { default as Dictionary } from './Dictionary';
 export { default as EmptyState } from './EmptyState';
-export { default as IconButton } from './IconButton';
 export { default as Key } from './Key';
 export { default as KeyMap } from './KeyMap';
 export { default as Loading } from './Loading';
